move preview-tabs to a folder
