+++ apps/web/app/(app)/talent/profile/preview/page.tsx
@@ -1,7 +1,7 @@
 import { ProfileCard } from './profile-card'
 import { me } from '@/lib/server/users'
 
-import { PreviewTabs } from './preview-tabs'
+import { PreviewTabs } from './tab-section/preview-tabs'
 import { ScrollCollapse } from './scroll-collapse'
 export default async function ProfilePage() {
   const user = await me()
@@ -10,7 +10,7 @@ export default async function ProfilePage() {
       <ScrollCollapse collapsedTitle={`${user.firstName} ${user.lastName}`}>
         <ProfileCard user={user} />
       </ScrollCollapse>
-      <PreviewTabs snapTarget="preview" />
+      <PreviewTabs />
     </div>
   )
 }
