+++ apps/web/app/(app)/talent/profile/preview/tab-section/preview-tabs.css
